Use create-react-app to initialize a new application.
Modify index.js to match the following.and remove extra files (App.js, App.css, App.test.js, logo.svg, serviceWorker.js).

Unfortunately, the entire application is in the same component. Refactor the code so that it consists of three new components: Header, Content, and Total. All data still resides in the App component, which passes the necessary data to each component using props. Header takes care of rendering the name of the course, Content renders the parts and their number of exercises and Total renders the total number of exercises.